Add global volume synced to localstorage
Positioning is a bit rough. Potentially you could get also sync the
volume of the players by listening for the `volumechange` event on each
of them. But the players in my browser don't seem to have a volume bar!

`accent-color: black` is to match the progress bars on the player in my
browser. Not sure how widely applicable that is, but the default color
is blue, and there is no other blue on the page.